+++ config/security.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'force_admin_two_factor' => env('FORCE_ADMIN_TWO_FACTOR', true),
+];